Trap ctrl-c in data download script
Make sure that data file is deleted if we press ctrl-c during the
download. Otherwise, we're left with a newer file that is smaller than
the one on the server. In theory, the -N option in wget should detect
the size discrepancy and download again when we next refresh the
data. But I've found that this doesn't seem to work for land registry
data.